// packages/core/src/domain/types.ts
export interface UserSettings {
  statsScale: 'relative' | 'absolute';
  simulationCount: number;
  timeLimit: '1min' | '5min' | '10min' | 'unlimited';
  version: number;
}